Create strategy file
